design: fab to root
