+++ src/components/layout.tsx
@@ -1,8 +1,6 @@
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import type { PropsWithChildren } from 'react';
 
-import AIFab from './ai-fab';
-import AIPopupWithRAG from './ai-popup-with-rag';
 // import { Footer } from './footer';
 import { Header } from './header';
 
@@ -10,32 +8,11 @@ export const Layout: React.FC<
   PropsWithChildren<{ disableHeaderHeight?: boolean }>
 > = ({ children, disableHeaderHeight }) => {
   const headerRef = useRef<HTMLElementTagNameMap['header']>(null);
-  const [isAIPopupOpen, setIsAIPopupOpen] = useState(false);
-
-  const toggleAIPopup = () => {
-    setIsAIPopupOpen(!isAIPopupOpen);
-  };
-
-  useEffect(() => {
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === 'Escape' && isAIPopupOpen) {
-        setIsAIPopupOpen(false);
-      }
-    };
-
-    document.addEventListener('keydown', handleKeyDown);
-    return () => {
-      document.removeEventListener('keydown', handleKeyDown);
-    };
-  }, [isAIPopupOpen]);
 
   return (
     <div className="min-h-screen w-full">
       <Header ref={headerRef} />
-      {/* AI FAB with integrated Function Calling */}
-      <AIFab isOpen={isAIPopupOpen} onClick={toggleAIPopup} />
-      <AIPopupWithRAG isOpen={isAIPopupOpen} />
-
+      
       <main
         className={`flex flex-col items-center mx-6 md:mx-12 lg:mx-24 xl:mx-48 ${disableHeaderHeight ? '' : 'pt-20'}`}
       >
